chore: add command, language and save managers
