+++ engine/game.py
@@ -1,74 +1,30 @@
-"""Core game loop."""
+"""Core game loop orchestrator."""
+
+from __future__ import annotations
 
 from pathlib import Path
-import re
-from dataclasses import dataclass
-from typing import Callable, cast
 
-import yaml
+from engine import io, parser, world, llm, integrity
 
-from engine import io, parser, world, llm, i18n, integrity
-
-
-@dataclass
-class StateChange:
-    state: str
-    message_key: str
-
-
-STATE_COMMANDS: dict[str, StateChange] = {
-    "destroy": StateChange("destroyed", "destroyed"),
-    "wear": StateChange("worn", "worn"),
-}
-
-
-@dataclass
-class ActionConfig:
-    trigger: str
-    item_in_inventory: bool
-    target_is_npc: bool
-    item_missing_key: str
-    target_missing_key: str
-    failure_key: str
-
-
-ACTION_COMMANDS: dict[str, ActionConfig] = {
-    "use": ActionConfig(
-        trigger="use",
-        item_in_inventory=False,
-        target_is_npc=False,
-        item_missing_key="use_failure",
-        target_missing_key="use_failure",
-        failure_key="use_failure",
-    ),
-    "show": ActionConfig(
-        trigger="show",
-        item_in_inventory=True,
-        target_is_npc=True,
-        item_missing_key="not_carrying",
-        target_missing_key="no_npc",
-        failure_key="use_failure",
-    ),
-}
+from .commands import CommandProcessor
+from .language import LanguageManager
+from .persistence import SaveManager
 
 
 class Game:
     def __init__(self, world_data_path: str, language: str, debug: bool = False) -> None:
         data_path = Path(world_data_path)
         self.data_dir = data_path.parent.parent
-        self.save_path = self.data_dir / "save.yaml"
-        generic_path = self.data_dir / "generic" / "world.yaml"
         self.debug = debug
+        self.save_manager = SaveManager(self.data_dir)
 
-        save_data: dict[str, object] = {}
-        if self.save_path.exists():
-            with open(self.save_path, encoding="utf-8") as fh:
-                save_data = yaml.safe_load(fh) or {}
-        self.language = str(save_data.get("language", language))
+        save_data = self.save_manager.load()
         self._show_intro = not save_data
-        lang_world_path = self.data_dir / self.language / "world.yaml"
+        self._language = str(save_data.get("language", language))
+        generic_path = self.data_dir / "generic" / "world.yaml"
+        lang_world_path = self.data_dir / self._language / "world.yaml"
 
-        warnings = integrity.check_translations(self.language, self.data_dir)
+        warnings = integrity.check_translations(self._language, self.data_dir)
         for msg in warnings:
             io.output(f"WARNING: {msg}")
 
@@ -84,342 +40,32 @@ class Game:
             raise SystemExit("Integrity check failed")
 
         if save_data:
-            self.world.load_state(self.save_path)
-            self.save_path.unlink()
+            self.world.load_state(self.save_manager.save_path)
+            self.save_manager.cleanup()
 
-        self.messages = i18n.load_messages(self.language)
-        self.commands = i18n.load_commands(self.language)
-        self.command_info = i18n.load_command_info()
-        self.command_keys = list(self.command_info.keys())
-        self.cmd_patterns: list[tuple[re.Pattern[str], str, str]] = []
-        self.reverse_cmds: dict[str, tuple[str, str]] = {}
-        self._build_cmd_patterns()
+        self.language_manager = LanguageManager(
+            self.data_dir, self._language, debug=debug
+        )
+        self.command_processor = CommandProcessor(
+            self.world,
+            self.language_manager,
+            self.save_manager,
+            self._check_end,
+            self._check_npc_event,
+            self.stop,
+            self._update_world,
+        )
         self.running = True
 
-    def run(self) -> None:
-        if self._show_intro and self.world.intro:
-            io.output(self.world.intro)
-        io.output(self.world.describe_current(self.messages))
-        self._check_npc_event()
-        self._check_end()
-        try:
-            while self.running:
-                raw = io.get_input()
-                raw = llm.interpret(raw)
-                raw = parser.parse(raw)
-                for pattern, cmd_key, _ in self.cmd_patterns:
-                    match = pattern.fullmatch(raw)
-                    if not match:
-                        continue
-                    groups = match.groupdict()
-                    handler = getattr(self, f"cmd_{cmd_key}", self.cmd_unknown)
-                    info = self.command_info.get(cmd_key, {})
-                    if info.get("arguments") == 2:
-                        a = groups.get("a", "").strip()
-                        b = groups.get("b", "").strip()
-                        two_handler = cast(Callable[[str, str], None], handler)
-                        two_handler(a, b)
-                    else:
-                        arg = groups.get("a", "") or groups.get("b", "") or ""
-                        one_handler = cast(Callable[[str], None], handler)
-                        one_handler(arg.strip())
-                    break
-                else:
-                    self.cmd_unknown(raw)
-        except (EOFError, KeyboardInterrupt):
-            io.output(self.messages["farewell"])
-        finally:
-            self._save_state()
+    @property
+    def language(self) -> str:
+        return self.language_manager.language
 
-    def _save_state(self) -> None:
-        data = self.world.to_state()
-        data["language"] = self.language
-        with open(self.save_path, "w", encoding="utf-8") as fh:
-            yaml.safe_dump(data, fh)
-
-    def _strip_suffix(self, arg: str, suffix: str) -> str:
-        if suffix and arg.endswith(f" {suffix}"):
-            return arg[: -len(suffix) - 1].strip()
-        return arg
-
-    def _find_item_id(self, name: str, *, in_inventory: bool = False) -> str | None:
-        if not name:
-            return None
-        name_cf = name.casefold()
-        if not in_inventory:
-            room = self.world.rooms[self.world.current]
-            for item_id in room.get("items", []):
-                names = self.world.items.get(item_id, {}).get("names", [])
-                if any(n.casefold() == name_cf for n in names):
-                    return item_id
-        for item_id in self.world.inventory:
-            names = self.world.items.get(item_id, {}).get("names", [])
-            if any(n.casefold() == name_cf for n in names):
-                return item_id
-        return None
-
-    def _find_npc_id(self, name: str) -> str | None:
-        if not name:
-            return None
-        name_cf = name.casefold()
-        for npc_id, npc in self.world.npcs.items():
-            names = npc.get("names", [])
-            if not any(n.casefold() == name_cf for n in names):
-                continue
-            if npc.get("meet", {}).get("location") != self.world.current:
-                return None
-            return npc_id
-        return None
-
-    def _state_command(self, cmd: str, item_name: str) -> None:
-        if not item_name:
-            self.cmd_unknown(item_name)
-            return
-        cfg = STATE_COMMANDS[cmd]
-        item_id = self._find_item_id(item_name, in_inventory=True)
-        if not item_id:
-            io.output(self.messages["not_carrying"])
-            self._check_end()
-            return
-        if not self.world.set_item_state(item_id, cfg.state):
-            io.output(self.messages["use_failure"])
-            self._check_end()
-            return
-        self.world.inventory.remove(item_id)
-        self.world.debug(f"inventory {self.world.inventory}")
-        io.output(self.messages[cfg.message_key].format(item=item_name))
-        self._check_end()
-
-    def _action_command(
-        self, cmd: str, item_name: str, target_name: str
-    ) -> None:
-        cfg = ACTION_COMMANDS[cmd]
-        if not item_name or not target_name:
-            self.cmd_unknown(cmd)
-            return
-        item_id = self._find_item_id(item_name, in_inventory=cfg.item_in_inventory)
-        if not item_id:
-            io.output(self.messages[cfg.item_missing_key])
-            self._check_end()
-            return
-        finder = self._find_npc_id if cfg.target_is_npc else self._find_item_id
-        target_id = finder(target_name)
-        if not target_id:
-            io.output(self.messages[cfg.target_missing_key])
-            self._check_end()
-            return
-        if self._execute_action(cfg.trigger, item_id, target_id):
-            self._check_end()
-            return
-        io.output(self.messages[cfg.failure_key])
-        self._check_end()
-
-    def _execute_action(
-        self, trigger: str, item_id: str, target_id: str | None = None
-    ) -> bool:
-        for action in self.world.actions:
-            if action.get("trigger") != trigger:
-                continue
-            if action.get("item") and action.get("item") != item_id:
-                continue
-            if action.get("target_item") and action.get("target_item") != target_id:
-                continue
-            if action.get("target_npc") and action.get("target_npc") != target_id:
-                continue
-            if not self.world.check_preconditions(action.get("preconditions")):
-                continue
-            effect = action.get("effect", {})
-            self.world.apply_effect(effect)
-            message = action.get("messages", {}).get("success")
-            if message:
-                io.output(message)
-            return True
-        return False
-
-    def describe_item(self, item_name: str) -> None:
-        if not item_name:
-            self.cmd_unknown(item_name)
-            return
-        item_id = self._find_item_id(item_name)
-        if not item_id:
-            io.output(self.messages["item_not_present"])
-            return
-        desc = self.world.describe_item(item_name)
-        if desc:
-            io.output(desc)
-        self._execute_action("examine", item_id)
-        self._check_end()
-
-    def cmd_quit(self, arg: str) -> None:
-        self._save_state()
-        io.output(self.messages["farewell"])
-        self.running = False
-
-    def cmd_inventory(self, arg: str) -> None:
-        io.output(self.world.describe_inventory(self.messages))
-
-    def cmd_take(self, arg: str) -> None:
-        if not arg:
-            self.cmd_unknown(arg)
-            return
-        item_name = arg
-        taken = self.world.take(item_name)
-        if taken:
-            io.output(self.messages["taken"].format(item=taken))
-        else:
-            io.output(self.messages["item_not_present"])
-        self._check_end()
-
-    def cmd_drop(self, arg: str) -> None:
-        if not arg:
-            self.cmd_unknown(arg)
-            return
-        item = arg
-        if self.world.drop(item):
-            io.output(self.messages["dropped"].format(item=item))
-        else:
-            io.output(self.messages["not_carrying"])
-        self._check_end()
-
-    def cmd_destroy(self, arg: str) -> None:
-        self._state_command("destroy", arg)
-
-    def cmd_wear(self, arg: str) -> None:
-        self._state_command("wear", arg)
-
-    def cmd_look(self, arg: str) -> None:
-        if arg:
-            self.cmd_unknown(arg)
-            return
-        io.output(self.world.describe_current(self.messages))
-
-    def cmd_examine(self, arg: str) -> None:
-        self.describe_item(arg)
-
-    def cmd_go(self, arg: str) -> None:
-        if not arg:
-            self.cmd_unknown(arg)
-            return
-        direction = arg
-        if self.world.can_move(direction) and self.world.move(direction):
-            io.output(self.world.describe_current(self.messages))
-            self._check_npc_event()
-        else:
-            io.output(self.messages["cannot_move"])
-        self._check_end()
-
-    def cmd_help(self, arg: str) -> None:
-        if not arg:
-            names: list[str] = []
-            for key in self.command_keys:
-                val = self.commands.get(key, [])
-                entries = val if isinstance(val, list) else [val]
-                first = entries[0]
-                names.append(first.split()[0])
-            io.output(self.messages["help"].format(commands=", ".join(names)))
-            return
-        cmd_info = self.reverse_cmds.get(arg)
-        if not cmd_info:
-            self.cmd_unknown(arg)
-            return
-        key, _ = cmd_info
-        entries = self.commands.get(key, [])
-        entries = entries if isinstance(entries, list) else [entries]
-        usages: list[str] = []
-        for entry in entries:
-            if self.command_info.get(key, {}).get("optional_arguments") and "$" not in entry:
-                continue
-            usage = entry.replace("$a", "<>").replace("$b", "<>")
-            usages.append(usage)
-        header = self.messages.get("help_usage", "Usage of \"{command}\" and synonyms:")
-        io.output(header.format(command=key) + "\n" + "\n".join(usages))
-
-    def cmd_language(self, arg: str) -> None:
-        if not arg:
-            self.cmd_unknown(arg)
-            return
-        language = arg.strip()
-        try:
-            messages = i18n.load_messages(language)
-            commands = i18n.load_commands(language)
-            world_path = self.data_dir / language / "world.yaml"
-            generic_path = self.data_dir / "generic" / "world.yaml"
-            new_world = world.World.from_files(
-                generic_path, world_path, debug=self.debug
-            )
-        except FileNotFoundError:
-            io.output(self.messages.get("language_unknown", "Unknown language"))
-            return
-        self._save_state()
-        new_world.load_state(self.save_path)
-        self.save_path.unlink()
+    def _update_world(self, new_world: world.World) -> None:
         self.world = new_world
-        self.language = language
-        self.messages = messages
-        self.commands = commands
-        self.cmd_patterns = []
-        self.reverse_cmds = {}
-        self._build_cmd_patterns()
-        io.output(self.messages["language_set"].format(language=language))
 
-    def cmd_show(self, item_name: str, npc_name: str) -> None:
-        self._action_command("show", item_name, npc_name)
-
-    def cmd_talk(self, arg: str) -> None:
-        if not arg:
-            self.cmd_unknown(arg)
-            return
-        npc_id = self._find_npc_id(arg)
-        if not npc_id:
-            io.output(self.messages["no_npc"])
-            return
-        npc = self.world.npcs[npc_id]
-        state = self.world.npc_state(npc_id)
-        talk_cfg = npc.get("states", {}).get(state, {})
-        text = talk_cfg.get("talk")
-        if text:
-            io.output(text)
-        else:
-            io.output(self.messages["no_npc"])
-        if state != "helped":
-            self.world.set_npc_state(npc_id, "helped")
-
-    def cmd_use(self, item_name: str, target_name: str) -> None:
-        self._action_command("use", item_name, target_name)
-
-    def _build_cmd_patterns(self) -> None:
-        for key in self.command_keys:
-            val = self.commands.get(key, [])
-            entries = val if isinstance(val, list) else [val]
-            for entry in entries:
-                pattern, base = self._compile_command(entry)
-                self.cmd_patterns.append((pattern, key, entry))
-                if base not in self.reverse_cmds:
-                    self.reverse_cmds[base] = (key, entry)
-        self.cmd_patterns.sort(key=lambda x: len(x[0].pattern), reverse=True)
-        self.reverse_cmds["language"] = ("language", "language")
-
-    def _compile_command(self, pattern: str) -> tuple[re.Pattern[str], str]:
-        tokens = pattern.split()
-        placeholder_positions = [i for i, t in enumerate(tokens) if t in ("$a", "$b")]
-        last_placeholder = placeholder_positions[-1] if placeholder_positions else -1
-        base = None
-        parts: list[str] = []
-        for idx, token in enumerate(tokens):
-            if token == "$a":
-                part = r"(?P<a>.+)" if idx == last_placeholder else r"(?P<a>.+?)"
-            elif token == "$b":
-                part = r"(?P<b>.+)" if idx == last_placeholder else r"(?P<b>.+?)"
-            else:
-                if base is None:
-                    base = token
-                part = re.escape(token)
-            parts.append(part)
-        regex = r"^" + r"\s+".join(parts) + r"$"
-        return re.compile(regex), base or pattern
-
-    def cmd_unknown(self, arg: str) -> None:
-        io.output(self.messages["unknown_command"])
+    def stop(self) -> None:
+        self.running = False
 
     def _check_end(self) -> None:
         ending = self.world.check_endings()
@@ -440,6 +86,24 @@ class Game:
                     io.output(text)
                 self.world.meet_npc(npc_id)
 
+    def run(self) -> None:
+        if self._show_intro and self.world.intro:
+            io.output(self.world.intro)
+        io.output(self.world.describe_current(self.language_manager.messages))
+        self._check_npc_event()
+        self._check_end()
+        try:
+            while self.running:
+                raw = io.get_input()
+                raw = llm.interpret(raw)
+                raw = parser.parse(raw)
+                self.command_processor.execute(raw)
+        except (EOFError, KeyboardInterrupt):
+            io.output(self.language_manager.messages["farewell"])
+        finally:
+            self.save_manager.save(self.world, self.language_manager.language)
+
 
 def run(world_data_path: str, language: str = "en", debug: bool = False) -> None:
     Game(world_data_path, language, debug=debug).run()
+
