+++ tests/test_help.py
@@ -5,14 +5,14 @@ def test_help_lists_commands(data_dir, monkeypatch):
     outputs: list[str] = []
     monkeypatch.setattr(io, "output", lambda text: outputs.append(text))
     g = game.Game(str(data_dir / "en" / "world.yaml"), "en")
-    g.cmd_help("")
+    g.command_processor.cmd_help("")
     names = []
-    for key in g.command_keys:
-        val = g.commands.get(key, [])
+    for key in g.command_processor.command_keys:
+        val = g.language_manager.commands.get(key, [])
         entries = val if isinstance(val, list) else [val]
         first = entries[0]
         names.append(first.split()[0])
-    expected = g.messages["help"].format(commands=", ".join(names))
+    expected = g.language_manager.messages["help"].format(commands=", ".join(names))
     assert outputs[-1] == expected
 
 
@@ -20,11 +20,11 @@ def test_help_lists_synonyms(data_dir, monkeypatch):
     outputs: list[str] = []
     monkeypatch.setattr(io, "output", lambda text: outputs.append(text))
     g = game.Game(str(data_dir / "en" / "world.yaml"), "en")
-    g.cmd_help("destroy")
-    entries = g.commands["destroy"]
+    g.command_processor.cmd_help("destroy")
+    entries = g.language_manager.commands["destroy"]
     usages = [e.replace("$a", "<>").replace("$b", "<>") for e in entries]
     expected = (
-        g.messages["help_usage"].format(command="destroy")
+        g.language_manager.messages["help_usage"].format(command="destroy")
         + "\n"
         + "\n".join(usages)
     )
@@ -35,9 +35,9 @@ def test_help_optional_argument(data_dir, monkeypatch):
     outputs: list[str] = []
     monkeypatch.setattr(io, "output", lambda text: outputs.append(text))
     g = game.Game(str(data_dir / "en" / "world.yaml"), "en")
-    g.cmd_help("help")
+    g.command_processor.cmd_help("help")
     expected = (
-        g.messages["help_usage"].format(command="help")
+        g.language_manager.messages["help_usage"].format(command="help")
         + "\nhelp <>\nh <>"
     )
     assert outputs[-1] == expected
