+++ tests/test_state_commands.py
@@ -9,6 +9,6 @@ def test_state_command_requires_existing_state(data_dir, monkeypatch, command):
     g = game.Game(str(data_dir / "en" / "world.yaml"), "en")
     assert g.world.move("Room 3")
     assert g.world.take("Sword")
-    getattr(g, f"cmd_{command}")("Sword")
-    assert outputs[-1] == g.messages["use_failure"]
+    getattr(g.command_processor, f"cmd_{command}")("Sword")
+    assert outputs[-1] == g.language_manager.messages["use_failure"]
     assert "sword" in g.world.inventory
